pyproject.toml
[project]
name = "r2r-rag-search-agent"
version = "0.1.0"
description = "MCP server for R2R (Retrieval-Augmented Generation) integration with Claude Desktop"
requires-python = ">=3.12"
dependencies = [
    "fastmcp==2.13.1",
    "r2r>=3.6.0",
    "google-genai>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.14.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "SIM", "UP", "RUF"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
